backend/requirements.txt
Flask
Flask-Cors
requests
pytz
python-dotenv
pymongo
gunicorn
pyswisseph==2.10.3.2
torch==2.9.0
accelerate==0.33.0
transformers==4.46.1
tokenizers>=0.20.0,<0.21
pydantic>=2.0.0
